refactor: move sampling http handler to internal/sampling/http
