+++ cmd/jaeger/internal/extension/remotesampling/extension.go
@@ -23,7 +23,7 @@ import (
 	"github.com/jaegertracing/jaeger/cmd/jaeger/internal/extension/jaegerstorage"
 	"github.com/jaegertracing/jaeger/internal/metrics/otelmetrics"
 	samplinggrpc "github.com/jaegertracing/jaeger/internal/sampling/grpc"
-	"github.com/jaegertracing/jaeger/pkg/clientcfg/clientcfghttp"
+	samplinghttp "github.com/jaegertracing/jaeger/internal/sampling/http"
 	"github.com/jaegertracing/jaeger/pkg/metrics"
 	"github.com/jaegertracing/jaeger/plugin/sampling/leaderelection"
 	"github.com/jaegertracing/jaeger/plugin/sampling/strategyprovider/adaptive"
@@ -229,8 +229,8 @@ func (ext *rsExtension) startAdaptiveStrategyProvider(host component.Host) error
 func (ext *rsExtension) startHTTPServer(ctx context.Context, host component.Host) error {
 	mf := otelmetrics.NewFactory(ext.telemetry.MeterProvider)
 	mf = mf.Namespace(metrics.NSOptions{Name: "jaeger_remote_sampling"})
-	handler := clientcfghttp.NewHTTPHandler(clientcfghttp.HTTPHandlerParams{
-		ConfigManager: &clientcfghttp.ConfigManager{
+	handler := samplinghttp.NewHandler(samplinghttp.HandlerParams{
+		ConfigManager: &samplinghttp.ConfigManager{
 			SamplingProvider: ext.strategyProvider,
 		},
 		MetricsFactory: mf,
